Fix counter_cache behavior for Rails 4.2
Previously we have just overridden the behavior of
`ActiveRecord::Persistence#destroy_row` to apply soft-delete
functionality. This does not work, since Rails depends on inheritance
to add extra functionality to `#destroy_row`. In Rails 4.2, this
includes `ActiveRecord::CounterCache#destroy_row` calling `super`, as
well as `ActiveRecord::Locking::Optimistic#destroy_row` also calling
`super`.

See also

https://github.com/rails/rails/pull/14735
https://github.com/rails/rails/pull/14765
