+++ lib/zombie_record/restorable.rb
@@ -51,18 +51,6 @@ module ZombieRecord
 
     private
 
-    # Override Rails' #destroy_row for soft-delete functionality
-    def destroy_row
-      time = current_time_from_proper_timezone
-
-      update_params = { deleted_at: time }
-      if self.class.column_names.include?("updated_at")
-        update_params[:updated_at] = time
-      end
-
-      update_columns(update_params)
-    end
-
     def restore_associated_records!
       self.class.reflect_on_all_associations.each do |association|
         # Only restore associations that are automatically destroyed alongside
